# api/app/openapi.yml
openapi: 3.1.0
info:
  title: MADB API
  version: 0.1.0
paths:
...
  /v1/tasks:
    post:
      tags:
      - tasks
      summary: Create Task
      operationId: create_task_v1_tasks_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - tasks
      summary: List Tasks
      operationId: list_tasks_v1_tasks_get
      parameters:
      - name: external_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: External Id
      - name: project_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
      - name: project_slug
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Slug
      - name: milestone_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Milestone Id
      - name: created_after
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created After
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskRead'
                title: Response List Tasks V1 Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/tasks/{task_id}:
    get:
      tags:
      - tasks
      summary: Get Task
      operationId: get_task_v1_tasks__task_id__get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - tasks
      summary: Update Task
      operationId: update_task_v1_tasks__task_id__patch
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AttachmentCreate:
      properties:
        name:
          type: string
          title: Name
        content_base64:
          type: string
          title: Content Base64
      type: object
      required:
      - name
      - content_base64
      title: AttachmentCreate
    AttachmentRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        task_id:
          type: string
          format: uuid
          title: Task Id
        name:
          type: string
          title: Name
        path:
          type: string
          title: Path
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - task_id
      - name
      - path
      - created_at
      title: AttachmentRead
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TaskCreate:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        milestone_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Milestone Id
        phase_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Phase Id
        parent_task_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Task Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
        persona_required:
          anyOf:
          - type: string
          - type: 'null'
          title: Persona Required
        acceptance_criteria:
          anyOf:
          - type: string
          - type: 'null'
          title: Acceptance Criteria
        effort_estimate:
          type: number
          title: Effort Estimate
          default: 0
        effort_spent:
          type: number
          title: Effort Spent
          default: 0
        priority_score:
          type: number
          title: Priority Score
          default: 0
        risk_level:
          type: string
          title: Risk Level
          default: low
        severity:
          type: string
          title: Severity
          default: minor
        status:
          type: string
          title: Status
          default: not_started
        external_id:
          anyOf:
          - type: string
          - type: 'null'
          title: External Id
        project_slug:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Slug
        milestone_slug:
          anyOf:
          - type: string
          - type: 'null'
          title: Milestone Slug
        parent_task_external_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Task External Id
        attachments:
          anyOf:
          - items:
              $ref: '#/components/schemas/AttachmentCreate'
            type: array
          - type: 'null'
          title: Attachments
        options:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Options
      type: object
      required:
      - project_id
      - title
      title: TaskCreate
    TaskPatch:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
        persona_required:
          anyOf:
          - type: string
          - type: 'null'
          title: Persona Required
        acceptance_criteria:
          anyOf:
          - type: string
          - type: 'null'
          title: Acceptance Criteria
        effort_estimate:
          anyOf:
          - type: number
          - type: 'null'
          title: Effort Estimate
        effort_spent:
          anyOf:
          - type: number
          - type: 'null'
          title: Effort Spent
        priority_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Priority Score
        risk_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk Level
        severity:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        phase_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Phase Id
        parent_task_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Task Id
        lock_version:
          type: integer
          title: Lock Version
          description: Current lock version for optimistic concurrency
      type: object
      required:
      - lock_version
      title: TaskPatch
    TaskRead:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        milestone_id:
          type: string
          format: uuid
          title: Milestone Id
        phase_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Phase Id
        parent_task_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Task Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
        persona_required:
          anyOf:
          - type: string
          - type: 'null'
          title: Persona Required
        acceptance_criteria:
          anyOf:
          - type: string
          - type: 'null'
          title: Acceptance Criteria
        effort_estimate:
          type: number
          title: Effort Estimate
          default: 0
        effort_spent:
          type: number
          title: Effort Spent
          default: 0
        priority_score:
          type: number
          title: Priority Score
          default: 0
        risk_level:
          type: string
          title: Risk Level
          default: low
        severity:
          type: string
          title: Severity
          default: minor
        status:
          type: string
          title: Status
          default: not_started
        external_id:
          anyOf:
          - type: string
          - type: 'null'
          title: External Id
        id:
          type: string
          format: uuid
          title: Id
        lock_version:
          type: integer
          title: Lock Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        attachments:
          anyOf:
          - items:
              $ref: '#/components/schemas/AttachmentRead'
            type: array
          - type: 'null'
          title: Attachments
        milestone_slug:
          anyOf:
          - type: string
          - type: 'null'
          title: Milestone Slug
      type: object
      required:
      - project_id
      - milestone_id
      - title
      - id
      - lock_version
      - created_at
      - updated_at
      title: TaskRead
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError